backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "amadeus>=12.0.0",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi-mail>=1.5.8",
    "fastapi[standard]>=0.121.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.4",
    "sqlmodel>=0.0.27",
]
